% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_FT}
\alias{calc_FT}
\title{Calculate FT
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity}
\usage{
calc_FT(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
TF
}
\description{
Calculate FT
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity
}
