% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_ionic_strength}
\alias{calc_ionic_strength}
\title{Ionic strength after Dickson (1990a); see Dickson et al. (2007)}
\usage{
calc_ionic_strength(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
Ionic strength
}
\description{
Ionic strength after Dickson (1990a); see Dickson et al. (2007)
}
