% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_K0}
\alias{calc_K0}
\title{Calculate K0}
\usage{
calc_K0(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
K0
}
\description{
Calculate K0
}
