% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_K1}
\alias{calc_K1}
\title{Calculate K1}
\usage{
calc_K1(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
K1
}
\description{
Calculate K1
}
