% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_K2}
\alias{calc_K2}
\title{Calculate K2}
\usage{
calc_K2(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
K2
}
\description{
Calculate K2
}
