% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KB}
\alias{calc_KB}
\title{Calculate KB}
\usage{
calc_KB(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
KB
}
\description{
Calculate KB
}
