% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KF}
\alias{calc_KF}
\title{Calculate KF}
\usage{
calc_KF(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KF
}
\description{
Calculate KF
}
