% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KP1}
\alias{calc_KP1}
\title{Calculate KP1}
\usage{
calc_KP1(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KP1
}
\description{
Calculate KP1
}
