% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KP3}
\alias{calc_KP3}
\title{Calculate KP3}
\usage{
calc_KP3(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KP3
}
\description{
Calculate KP3
}
