% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_Ks}
\alias{calc_Ks}
\title{Calculate multiple equilibrium constants for carbon}
\usage{
calc_Ks(
  ks = NULL,
  TC = 25,
  S = 35,
  Mg = 0.0528171,
  Ca = 0.0102821,
  P = NULL,
  method = "MyAMI"
)
}
\arguments{
\item{ks}{character vectors of Ks to be calculated e.g., c("K0", "K1")}

\item{TC}{Temperature (Celsius)}

\item{S}{Salinity (PSU)}

\item{Mg}{Mg concentration in mol/kgsw. If None, modern is assumed (0.0528171). Should be the average Mg concentration in seawater - a salinity correction is then applied to calculate the Mg concentration in the sample.}

\item{Ca}{Ca concentration in mol/kgsw. If None, modern is assumed (0.0102821). Should be the average Ca concentration in seawater - a salinity correction is then applied to calculate the Mg concentration in the sample.}

\item{P}{Pressure (Bar) (optional)}

\item{method}{Options: \code{R_Polynomial}, \code{MyAMI_Polynomial} , \code{MyAMI} (defaults to "MyAMI").}
}
\value{
Data.frame of specified Ks at the given conditions
}
\description{
Wrapper to calculate multiple stoichiometric equilibrium constants at given temperature, salinity and pressure.
}
\author{
Dennis Mayk
}
