% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KSi}
\alias{calc_KSi}
\title{Calculate KSi}
\usage{
calc_KSi(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
KSi
}
\description{
Calculate KSi
}
