% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KW}
\alias{calc_KW}
\title{Calculate KW}
\usage{
calc_KW(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KW
}
\description{
Calculate KW
}
