% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_pc}
\alias{calc_pc}
\title{Calculate pressure correction factor for Ks}
\usage{
calc_pc(coefficients, temp_c, p_bar)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{p_bar}{Pressure (Bar)}
}
\value{
Pressure correction factor
}
\description{
Calculate pressure correction factor for Ks
}
\references{
From Millero et al. (2007, doi:10.1021/cr0503557), Eqns 38-40
}
