% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_pressure_correction}
\alias{calc_pressure_correction}
\title{Calculate pressure correction factor}
\usage{
calc_pressure_correction(k, TC, P)
}
\arguments{
\item{k}{K to be calculated}

\item{TC}{Temperature (Celsius)}

\item{P}{Pressure (Bar)}
}
\description{
Calculate pressure correction factors for a specified equilibrium constant
}
\author{
Dennis Mayk
}
