% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_ST}
\alias{calc_ST}
\title{Calculate ST
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity#'}
\usage{
calc_ST(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
TS
}
\description{
Calculate ST
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity#'
}
