% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_sulphate}
\alias{calc_sulphate}
\title{Calculate sulphate}
\usage{
calc_sulphate(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
sulphate
}
\description{
Calculate sulphate
}
\references{
From Dickson et al., 2007, Table 2, Note: Sal / 1.80655 = Chlorinity
}
