% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_TF}
\alias{calc_TF}
\title{Calculate TF
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity}
\usage{
calc_TF(S)
}
\arguments{
\item{S}{Salinity}
}
\value{
TF
}
\description{
Calculate TF
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity
}
