% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_TS}
\alias{calc_TS}
\title{Calculate TS
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity#'}
\usage{
calc_TS(S)
}
\arguments{
\item{S}{Salinity}
}
\value{
TS
}
\description{
Calculate TS
From Dickson et al., 2007, Table 2
Note: Sal / 1.80655 = Chlorinity#'
}
