% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_Istr}
\alias{fn_Istr}
\title{Ionic strength after Dickson (1990a); see Dickson et al. (2007)}
\usage{
fn_Istr(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
Ionic strength
}
\description{
Ionic strength after Dickson (1990a); see Dickson et al. (2007)
}
