% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_K0}
\alias{fn_K0}
\title{Calculate K0}
\usage{
fn_K0(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
K0
}
\description{
Calculate K0
}
