% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_K1}
\alias{fn_K1}
\title{Calculate K1}
\usage{
fn_K1(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
K1
}
\description{
Calculate K1
}
