% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KF}
\alias{fn_KF}
\title{Calculate KF}
\usage{
fn_KF(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KF
}
\description{
Calculate KF
}
