% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KP1}
\alias{fn_KP1}
\title{Calculate KP1}
\usage{
fn_KP1(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KP1
}
\description{
Calculate KP1
}
