% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KP3}
\alias{fn_KP3}
\title{Calculate KP3}
\usage{
fn_KP3(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
KP3
}
\description{
Calculate KP3
}
