% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KS}
\alias{fn_KS}
\title{Calculate KS}
\usage{
fn_KS(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KS
}
\description{
Calculate KS
}
