% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_Ksp}
\alias{fn_Ksp}
\title{Calculate Ksp}
\usage{
fn_Ksp(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
Ksp
}
\description{
Calculate Ksp
}
