% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KW}
\alias{fn_KW}
\title{Calculate KW}
\usage{
fn_KW(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KW
}
\description{
Calculate KW
}
