% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kgen-package.R
\docType{package}
\name{Kgen-package}
\alias{Kgen}
\alias{Kgen-package}
\title{Kgen: Parameters for Consistently Calculating Ks for Carbon Calculation}
\description{
Coefficients for consistently calculating and pressure correcting Ks for carbon calculation.
}
\author{
\strong{Maintainer}: Dennis Mayk \email{dm807@cantab.ac.uk} (\href{https://orcid.org/0000-0002-5017-1495}{ORCID})

}
\keyword{internal}
