% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{kgen_poly}
\alias{kgen_poly}
\title{Kgen R polynomial function}
\usage{
kgen_poly(S, TK, Mg, Ca)
}
\arguments{
\item{S}{Salinity (PSU)}

\item{TK}{Temperature (Kelvin)}

\item{Mg}{Mg concentration in mol/kgsw. If None, modern is assumed (0.0528171). Should be the average Mg concentration in seawater - a salinity correction is then applied to calculate the Mg concentration in the sample.}

\item{Ca}{Ca concentration in mol/kgsw. If None, modern is assumed (0.0102821). Should be the average Ca concentration in seawater - a salinity correction is then applied to calculate the Mg concentration in the sample.}
}
\description{
Kgen R polynomial function
}
