% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymyami.R
\name{miniconda_conda}
\alias{miniconda_conda}
\title{Check if miniconda is installed}
\usage{
miniconda_conda(path = miniconda_path())
}
\arguments{
\item{path}{Path to miniconda}
}
\description{
Check if miniconda is installed
}
