% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_fluorine}
\alias{calc_fluorine}
\title{Calculate fluorine}
\usage{
calc_fluorine(sal)
}
\arguments{
\item{sal}{Salinity}
}
\value{
fluorine
}
\description{
Calculate fluorine
}
\references{
From Dickson et al., 2007, Table 2, Note: Sal / 1.80655 = Chlorinity
}
