% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_K}
\alias{calc_K}
\title{Calculate a specified stoichiometric equilibrium constants at given temperature, salinity and pressure.}
\usage{
calc_K(k, TC = 25, S = 35, P = NA)
}
\arguments{
\item{k}{K to be calculated}

\item{TC}{Temperature (Celsius)}

\item{S}{Salinity (PSU)}

\item{P}{Pressure (Bar) (optional)}
}
\value{
Specified K at the given conditions
}
\description{
Calculate a specified stoichiometric equilibrium constants at given temperature, salinity and pressure.
}
\examples{
calc_K("K0", 25, 35)
}
