% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KP2}
\alias{calc_KP2}
\title{Calculate KP2}
\usage{
calc_KP2(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KP2
}
\description{
Calculate KP2
}
