% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_Ks}
\alias{calc_Ks}
\title{Calculate a specified stoichiometric equilibrium constants at given temperature, salinity and pressure.}
\usage{
calc_Ks(TC = 25, S = 35, P = NA, K_list)
}
\arguments{
\item{TC}{Temperature (Celsius)}

\item{S}{Salinity (PSU)}

\item{P}{Pressure (Bar) (optional)}

\item{K_list}{List of Ks to be calculated e.g., list("K0", "K1")}
}
\value{
Dataframe of specified Ks at the given conditions
}
\description{
Calculate a specified stoichiometric equilibrium constants at given temperature, salinity and pressure.
}
\examples{
calc_Ks(25, 35, K_list = c("K0", "K1"))
}
