% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_KSi}
\alias{calc_KSi}
\title{Calculate KSi}
\usage{
calc_KSi(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
KSi
}
\description{
Calculate KSi
}
