% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{calc_Ksp}
\alias{calc_Ksp}
\title{Calculate Ksp}
\usage{
calc_Ksp(coefficients, temp_c, sal)
}
\arguments{
\item{coefficients}{Coefficients for K calculation}

\item{temp_c}{Temperature (Celcius)}

\item{sal}{Salinity}
}
\value{
Ksp
}
\description{
Calculate Ksp
}
