% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc_pressure_correction}
\alias{calc_pressure_correction}
\title{Calculate pressure correction factor}
\usage{
calc_pressure_correction(k, temp_c, p_bar)
}
\arguments{
\item{k}{K to be calculated}

\item{temp_c}{Temperature (Celcius)}

\item{p_bar}{Pressure (Bar)}
}
\value{
pressure correction factor
}
\description{
Calculate pressure correction factor for a specified equilibrium constant.
}
\author{
Dennis Mayk
}
