% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_K2}
\alias{fn_K2}
\title{Calculate K2}
\usage{
fn_K2(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
K2
}
\description{
Calculate K2
}
