% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KB}
\alias{fn_KB}
\title{Calculate KB}
\usage{
fn_KB(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KB
}
\description{
Calculate KB
}
