% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KP1}
\alias{fn_KP1}
\title{Calculate KP1}
\usage{
fn_KP1(p, temp_k, sal)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_k}{Temperature (Kelvin)}

\item{sal}{Salinity}
}
\value{
KP1
}
\description{
Calculate KP1
}
