% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_KP3}
\alias{fn_KP3}
\title{Calculate KP3}
\usage{
fn_KP3(p, TK, S)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{TK}{Temperature (Kelvin)}

\item{S}{Salinity}
}
\value{
KP3
}
\description{
Calculate KP3
}
