% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_functions.R
\name{fn_pc}
\alias{fn_pc}
\title{Calculate pressure correction factor for Ks
From Millero et al. (2007, doi:10.1021/cr0503557)
Eqns 38-40}
\usage{
fn_pc(p, temp_c, p_bar)
}
\arguments{
\item{p}{Parameters for K calculation}

\item{temp_c}{Temperature (Celsius)}

\item{p_bar}{Pressure}
}
\value{
Pressure correction factor
}
\description{
Calculate pressure correction factor for Ks
From Millero et al. (2007, doi:10.1021/cr0503557)
Eqns 38-40
}
