% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kgen-package.R
\docType{package}
\name{kgen-package}
\alias{kgen}
\alias{kgen-package}
\title{kgen: A Tool for Calculating Stoichiometric Equilibrium Constants (Ks) for Seawater}
\description{
A unified software package simultaneously implemented in 'Python', 'R', and 'Matlab' providing a uniform and internally-consistent way of calculating stoichiometric equilibrium constants in modern and palaeo seawater as a function of temperature, salinity, pressure and the concentration of magnesium, calcium, sulphate, and fluorine.
}
\author{
\strong{Maintainer}: Dennis Mayk \email{dm807@cantab.ac.uk} (\href{https://orcid.org/0000-0002-5017-1495}{ORCID})

}
\keyword{internal}
