% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{kgen_poly}
\alias{kgen_poly}
\title{Kgen R polynomial function}
\usage{
kgen_poly(sal, temp_c, magnesium = 0.0528171, calcium = 0.0102821)
}
\arguments{
\item{sal}{Salinity}

\item{temp_c}{Temperature (Celcius)}

\item{magnesium}{magnesium concentration in mol/kgsw. If None, modern is assumed (0.0528171). Should be the average magnesium concentration in seawater - a salinity correction is then applied to calculate the magnesium concentration in the sample.}

\item{calcium}{calcium concentration in mol/kgsw. If None, modern is assumed (0.0102821). Should be the average calcium concentration in seawater - a salinity correction is then applied to calculate the magnesium concentration in the sample.}
}
\description{
Kgen R polynomial function
}
\author{
Dennis Mayk
}
